refactor(wip): use generics
