+++ okex-client/src/lib.rs
@@ -1,3 +1,3 @@
-mod client;
+mod client_v1;
 
-pub use client::*;
+pub use client_v1::*;
